+++ src/common/logger.ts
@@ -1,19 +1,21 @@
 import { createLogger, format, transports } from 'winston';
 
+const fileFormat = format.combine(format.uncolorize(), format.json());
+
 const logger = createLogger({
   level: 'silly',
-  format: format.combine(format.cli()),
+  format: format.combine(format.colorize(), format.cli()),
   transports: [
     new transports.Console(),
     new transports.File({
       filename: 'error.log',
       level: 'error',
-      format: format.combine(format.cli(), format.json()),
+      format: fileFormat,
     }),
     new transports.File({
       filename: 'info.log',
       level: 'info',
-      format: format.combine(format.cli(), format.json()),
+      format: fileFormat,
     }),
   ],
 });
