+++ src/middleware/error-handling.ts
@@ -1,6 +1,6 @@
 import { StatusCodes, getReasonPhrase } from 'http-status-codes';
 import { NextFunction, Request, Response } from 'express';
-import { logger } from './logger';
+import { logger } from '../common/logger';
 
 const errorHandler = async (
   err: Error,
@@ -8,7 +8,7 @@ const errorHandler = async (
   res: Response,
   _next: NextFunction
 ): Promise<void> => {
-  logger.error(`failure: ${err}`);
+  logger.error(`unhandled error: ${err.name}: ${err.message}`);
   res
     .status(StatusCodes.INTERNAL_SERVER_ERROR)
     .send(getReasonPhrase(StatusCodes.INTERNAL_SERVER_ERROR));
